CNN 과제/[0227]CNN_SeunghyeonLEE/requirements.txt
matplotlib
numpy
pandas
Pillow
PyYAML
scikit-learn
seaborn
torch
torchvision
